backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-ollama
langchain-core
langchain-chroma
sqlalchemy
python-dotenv
pymysql